Refactoring, some jsdocs
