--- packages/cubejs-server-core/src/core/DriverDependencies.js
@@ -1,28 +1,0 @@
-module.exports = {
-  postgres: '@cubejs-backend/postgres-driver',
-  mysql: '@cubejs-backend/mysql-driver',
-  mysqlauroraserverless: '@cubejs-backend/mysql-aurora-serverless-driver',
-  mssql: '@cubejs-backend/mssql-driver',
-  athena: '@cubejs-backend/athena-driver',
-  jdbc: '@cubejs-backend/jdbc-driver',
-  mongobi: '@cubejs-backend/mongobi-driver',
-  bigquery: '@cubejs-backend/bigquery-driver',
-  redshift: '@cubejs-backend/redshift-driver',
-  clickhouse: '@cubejs-backend/clickhouse-driver',
-  firebolt: '@cubejs-backend/firebolt-driver',
-  hive: '@cubejs-backend/hive-driver',
-  snowflake: '@cubejs-backend/snowflake-driver',
-  prestodb: '@cubejs-backend/prestodb-driver',
-  oracle: '@cubejs-backend/oracle-driver',
-  sqlite: '@cubejs-backend/sqlite-driver',
-  awselasticsearch: '@cubejs-backend/elasticsearch-driver',
-  elasticsearch: '@cubejs-backend/elasticsearch-driver',
-  dremio: '@cubejs-backend/dremio-driver',
-  druid: '@cubejs-backend/druid-driver',
-  cubestore: '@cubejs-backend/cubestore-driver',
-  ksql: '@cubejs-backend/ksql-driver',
-  questdb: '@cubejs-backend/questdb-driver',
-  materialize: '@cubejs-backend/materialize-driver',
-  // List for JDBC drivers
-  'databricks-jdbc': '@cubejs-backend/databricks-jdbc-driver',
-};
